Moved user state to seperate module
