+++ src/models/user.ts
@@ -1,0 +1,4 @@
+export type UserState = {
+  email: string,
+  token: string
+}
